XMLHttpRequest/open-url-bogus.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: open() - bogus URLs</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      function url(url) {
        test(function() {
          var client = new XMLHttpRequest()
          assert_throws("SYNTAX_ERR", function() { client.open("GET", url) })
        }, document.title + " (" + url + ")")
      }
      url("http:")
      url("http://a a/")
    </script>
  </body>
</html>
